+++ apps/store/src/components/ProductPage/ProductPage.types.ts
@@ -1,8 +1,8 @@
 import type { ProductData } from '@/components/ProductData/ProductData.types'
+import { type TrustpilotData } from '@/features/memberReviews/trustpilot.types'
 import { Template } from '@/services/PriceCalculator/PriceCalculator.types'
 import type { AverageRating, ReviewComments } from '@/services/productReviews/productReviews.types'
 import { ProductStory, StoryblokPageProps } from '@/services/storyblok/storyblok'
-import { type TrustpilotData } from '@/services/trustpilot/trustpilot.types'
 
 export type ProductPageProps = StoryblokPageProps & {
   story: ProductStory
