+++ apps/store/src/components/ProductReviews/ProductReviews.tsx
@@ -1,12 +1,12 @@
 import styled from '@emotion/styled'
 import React, { useState } from 'react'
 import { Space, theme } from 'ui'
+import { TrustpilotWidget } from '@/features/memberReviews/TruspilotWidget'
+import { useTrustpilotData } from '@/features/memberReviews/trustpilot'
+import { type Review as CompanyReview } from '@/features/memberReviews/trustpilot.types'
 import { getReviewsDistribution } from '@/services/productReviews/getReviewsDistribution'
 import { MAX_SCORE } from '@/services/productReviews/productReviews.constants'
 import type { Review as ProductReview } from '@/services/productReviews/productReviews.types'
-import { TrustpilotWidget } from '@/services/trustpilot/TruspilotWidget'
-import { useTrustpilotData } from '@/services/trustpilot/trustpilot'
-import { type Review as CompanyReview } from '@/services/trustpilot/trustpilot.types'
 import { useProductPageContext } from '../ProductPage/ProductPageContext'
 import { AverageRating } from './AverageRating'
 import type { Score, Review, ReviewsDistribution } from './ProductReviews.types'
